server/requirements.txt
# FastAPI requirements
fastapi[standard]
pydub
python-dotenv
redis
ruff
